Fix backtrace on fatal error on FreeBSD Aarch64
The BSD implementation for printing the backtraces on fatal error
reports from the JVM did not properly handle stack frames for compiled
java code on Aarch64.

This patch fixes that for FreeBSD by replicating the code from the Linux
implementation. I left it as it was for other BSD's, as I don't know if
the same issue is relevant for them.

This work was sponsored by: The FreeBSD Foundation